api/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
httpx>=0.27.0
asyncpg>=0.29.0
pydantic>=2.7.0
pydantic-settings>=2.3.0
python-dotenv>=1.0.0
yfinance>=0.2.40
PyJWT[crypto]>=2.8.0
cryptography>=42.0.0
